+++ ShowTweet/CHANGELOG.md
@@ -1,5 +1,8 @@
 #Changelog
 ---
+##1.1
+- Added checkmark for verified accounts
+- Fixed module not working due to Twitter changing things
 
 #1.0
 - Initial release